Restart LIFX client when network comes back up
This refreshes the state when the computer wakes up from sleep.
Additionally, upon the first network state change it will learn the home
network and not try to restart the LIFX client on other networks.
There’s no UI to change this so if you want to change home network, you
need to use the defaults command and manually delete the old value.

RHSNetworkMonitor is a wrapper around the System Configuration framework
that lets you get notified of IPv4 network interface changes with a
simple callback block.